feat: Primary Actions
BREAKING CHANGE:  Previous version was extending `dw-select`. This version doesn't extend it anymore, in fact it uses dw-select as a local DOM. During this change, we have considered maximum properties of dw-select which might be used by the users and they are being proxied through the same named properties. But, still we don't proxy each and every property. So, it may break your integration.